began work on @octokit stuff instead of octokit
